+++ Sources/VCore/Extensions/SwiftUI/NavigationPath.Methods.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 19.06.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - Navigation Path Methods
@@ -32,5 +30,3 @@ extension NavigationPath {
         removeLast(count)
     }
 }
-
-#endif