+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/CoordinatingNavigationStack.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 19.06.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - Coordinating Navigation Stack
@@ -70,5 +68,3 @@ public struct CoordinatingNavigationStack<Root>: View where Root: View {
             .environment(\.navigationStackCoordinator, navigationStackCoordinator)
     }
 }
-
-#endif