+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/EnvironmentValues.NavigationStackCoordinator.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 07.08.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - Environment Values Navigation Stack Coordinator
@@ -23,5 +21,3 @@ extension EnvironmentValues {
 struct NavigationStackCoordinatorEnvironmentKey: EnvironmentKey {
     static var defaultValue: NavigationStackCoordinator? = nil
 }
-
-#endif