+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/NavigationStackCoordinator.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 07.08.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - Navigation Stack Coordinator
@@ -25,5 +23,3 @@ public final class NavigationStackCoordinator: ObservableObject {
         self.path = path
     }
 }
-
-#endif