+++ Sources/VCore/Models/Text Line Type/TextLineLimitType.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 14.08.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - View Line Limit
@@ -87,5 +85,3 @@ private struct TextLineLimitViewModifier: ViewModifier {
         }
     }
 }
-
-#endif