+++ Sources/VCore/Models/Text Line Type/TextLineType.swift
@@ -5,8 +5,6 @@
 //  Created by Vakhtang Kontridze on 17.08.22.
 //
 
-#if !os(macOS) // TODO: Support macOS on release of 13.0
-
 import SwiftUI
 
 // MARK: - Text Line Type
@@ -113,5 +111,3 @@ enum _TextLineType {
     case singleLine
     case multiLine(alignment: TextAlignment, textLineLimitType: TextLineLimitType)
 }
-
-#endif